+++ vite.config.js
@@ -34,8 +34,7 @@ export default defineConfig({
       output: {
         globals: {
           react: 'React',
-          'react-dom': 'ReactDOM',
-          'prop-types': 'PropTypes'
+          'react-dom': 'ReactDOM'
         }
       }
     }