+++ CHANGELOG.md
@@ -28,3 +28,6 @@ When this package was imported from npm into a vite based project, it complained
 ## 2025/12/10 - Version 1.0.7 - ITA
 - Change all default exports to named exports. This is to improve tree-shaking in consumers of this package that use modern bundlers like Vite and WebPack.
 - useDialogBox function to throw an error if used in a component not wrapped inside a DialogBoxProvider.
+
+## 2025/12/10 - Version 1.0.8 - ITA
+- Style the DialogBox into a 'blank canvas', leaving developers to only worry about styling the contents displayed inside the DialogBox.